% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{block_average}
\alias{block_average}
\title{Perform mean-shift detection}
\usage{
block_average(x, width = 14, ccB = 2, idx.outlier = NULL)
}
\arguments{
\item{x}{Integer. A vector of integers.}

\item{width}{Integer. Size of window}

\item{ccB}{Numeric. Threshold level for mean-shift detection.}

\item{idx.outlier}{Integer. A vector of series indexes specifying outliers to
control for before applying algorithm.}
}
\value{
Index vector of detected outliers
}
\description{
Perform mean-shift detection using a rolling window
}
\details{
Two adjacent blocks (each of size \code{width}) of the series create
a rolling window. At each index, we test for statistically significant
differences in the mean of the two series blocks.
}
\seealso{
Algorithm details on Hammer at ~/Documents/aF_products/ts_anomaly
}

