% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{count_dist_outlier}
\alias{count_dist_outlier}
\title{Perform parametric outlier detection for count data}
\usage{
count_dist_outlier(x, sigma = 0.01)
}
\arguments{
\item{x}{Integer. A vector of integers.}

\item{sigma}{Numeric. The significance level for outlier detection.}
}
\value{
Index vector of detected outliers
}
\description{
This function fits three different parametric count data distributions: poisson,
negative binomial, and geometric. Selection between the three is carried out
by choosing which returned the largest maximum likelihood.
}

