% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{detect_meanshift}
\alias{detect_meanshift}
\title{Perform jump detection and mean-shift detection}
\usage{
detect_meanshift(x, width = 7, ccB = 2, p = 0.01, idx.outliers = NULL)
}
\arguments{
\item{x}{Integer. A vector of integers.}

\item{width}{Integer. Size of window}

\item{ccB}{Numeric. Threshold level for mean-shift detection.}

\item{p}{Numeric. Threshold level for jump detection.}

\item{idx.outliers}{Integer. A vector of series indexes specifying outliers to
control for before applying algorithm.}
}
\value{
A list containing both outlier indicies from \code{\link{seq_average}}
and \code{\link{block_average}}.
}
\description{
Perform jump detection and mean-shift detection
}
\seealso{
Algorithm details on Hammer at ~/Documents/aF_products/ts_anomaly
}

