% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{detect_varshift}
\alias{detect_varshift}
\title{Performs variance shift detection}
\usage{
detect_varshift(x, width = 7, var.diff = FALSE, idx.outliers = NULL)
}
\arguments{
\item{x}{Integer. A vector of integers.}

\item{width}{Integer. Value need to control for outliers passed to function by
\code{idx.outliers}.}

\item{var.diff}{Logical. Should the series be differenced before algorithm is
run.}

\item{idx.outliers}{Integer. A vector of series indexes specifying outliers to
control for before applying algorithm.}
}
\value{
A list containing index of variance shifts.
}
\description{
Performs variance shift detection
}

