% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{iso_forest}
\alias{iso_forest}
\title{Perform Isolation Forest outlier detection}
\usage{
iso_forest(x, STL = TRUE, ccF = 2.5)
}
\arguments{
\item{x}{Integer. A vector of integers.}

\item{STL}{Logical. Should the series be seasonally detrended before isolation
forest is run.}

\item{ccF}{Numeric. The significance level for outlier detection. See Details.}
}
\value{
Index vector of detected outliers
}
\description{
Run an isolation forest
}
\details{
The result of the isolation forest assigns an anomaly score for each
observations. These observations are transformed by a z-score and then those
z-scores above \code{ccF} are considered outliers.
}

