% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{merge_anomaly}
\alias{merge_anomaly}
\title{Translates anomalies in R lists to json}
\usage{
merge_anomaly(lst)
}
\arguments{
\item{lst}{List. Output from \code{\link{segment_anomaly}} or
\code{\link{anomaly_detection}}}
}
\value{
A list containing index of variance shifts.
}
\description{
Translates anomalies in R lists to json
}

