% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{oplot}
\alias{oplot}
\title{Plot detected anomalies}
\usage{
oplot(obj, ...)
}
\arguments{
\item{obj}{S3 object created by call to \code{\link{anomaly_detection}} or
\code{\link{segment_anomaly}}.}

\item{...}{other paramaters to pass to \code{plot}}
}
\description{
Plot detected anomalies
}

