% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{regroup_segments}
\alias{regroup_segments}
\title{Merge list of sequences that contain too few observations}
\usage{
regroup_segments(lst, mindays = 30, combine = c("last", "next"))
}
\arguments{
\item{lst}{List. A list of sequences.}

\item{mindays}{Integer. The lower threshhold. Sequences with fewer than the
threshold are merged withand adjacent sequence in the list.}

\item{combine}{Character. Combine offending list with preceding ('last') or
the following ('next') sequence in the list. This doesn't apply to the first
or last sequence in the list for obvious reasons.}
}
\value{
A list of integer sequences
}
\description{
Takes a list of sequences and merges adjacent sequences if one or more has
fewer than a specified minimum number of entries.
}

