% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{segment}
\alias{segment}
\title{Segment series into adjacent regimes}
\usage{
segment(Y, K = 5L, alpha = 0.05, exact = FALSE)
}
\arguments{
\item{Y}{Numeric. Vector to be segmented}

\item{K}{Integer. Number of segments. Defaults to 5.}

\item{alpha}{Numeric. Critical value for F-statistic. Defaults to 0.05.}

\item{exact}{Logical. Skip reduction of time series. Defaults to FALSE.}
}
\value{
Vector of integers specifying segmentation points of the series.
}
\description{
Segment series into adjacent regimes
}

