% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{segment_anomaly}
\alias{segment_anomaly}
\title{Perform Anomaly Detection with Series Segmentation}
\usage{
segment_anomaly(x, start_date = NULL, end_date = NULL, type = c("outlier",
  "meanshift", "varshift"), ctr.outlier = TRUE, alpha = 0.01,
  var.diff = FALSE, STL = FALSE, sigma = 0.01, ccF = 2.5, width = 7,
  ccB = 2, p = 0.01, consensus = TRUE, json = TRUE)
}
\arguments{
\item{x}{Integer. A vector of integers.}

\item{start_date}{Character. Date of first observation in the series \code{x}.}

\item{end_date}{Character. Date of last observation in the series \code{x}.}

\item{type}{Character. Select which anomaly test to run. Defaults to all.}

\item{ctr.outlier}{Logical. Should we pass detected outliers to meanshift and
jump detection algorithms}

\item{alpha}{Numeric. Significance level for segmentation process}

\item{var.diff}{Logical. Should the series be differenced before algorithm is
run.}

\item{STL}{Logical. Should the series be seasonally detrended before isolation
forest is run.}

\item{sigma}{Numeric. The significance level for outlier detection.}

\item{ccF}{Numeric. The significance level for outlier detection. See Details.}

\item{width}{Integer. Size of window}

\item{ccB}{Numeric. Threshold level for mean-shift detection.}

\item{p}{Numeric. Threshold level for jump detection.}

\item{consensus}{Logical. If \code{TRUE}, report anomalies only when all
algorithms agree.}

\item{json}{Logical. If \code{TRUE}, yield output in json.}
}
\value{
A list the \code{x}, \code{start_date}, \code{end_date}, and the output
from \code{\link{detect_outliers}}, \code{\link{detect_meanshift}}, and
\code{\link{detect_varshift}}.
}
\description{
Perform Anomaly Detection with Series Segmentation
}

