% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{segment_index}
\alias{segment_index}
\title{Fill in vector sequence}
\usage{
segment_index(x, overlap = 7)
}
\arguments{
\item{x}{Integer. A vector of integers.}

\item{overlap}{Integer. How much overlap to add to each sequence}
}
\value{
A list of integer vectors
}
\description{
Takes a vector of integers provided by segment.cpp and returns a list of
integer sequences. The start/end points of each sequence are the values of 
the inputed integer vector.
of integers
}

