% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnomaly.R
\name{seq_average}
\alias{seq_average}
\title{Perform jump detection}
\usage{
seq_average(x, width = 14, p = 0.01, idx.outlier = NULL)
}
\arguments{
\item{x}{Integer. A vector of integers.}

\item{width}{Integer. Size of window}

\item{p}{Numeric. Threshold level for jump detection.}

\item{idx.outlier}{Integer. A vector of series indexes specifying outliers to
control for before applying algorithm.}
}
\value{
Index vector of detected outliers
}
\description{
Perform jump detection using a sequential average algorithm
}
\seealso{
Algorithm details on Hammer at ~/Documents/aF_products/ts_anomaly
}

